enhancement(Dashboard/IncomingContributions): adapt for non-hosts
