+++ components/dashboard/sections/contributions/IncomingContributions.tsx
@@ -1,132 +1,14 @@
-import React, { useContext } from 'react';
-import { useQuery } from '@apollo/client';
-import { FormattedMessage } from 'react-intl';
-import type { z } from 'zod';
+import React from 'react';
 
-import type { FilterComponentConfigs, FiltersToVariables } from '../../../../lib/filters/filter-types';
-import { API_V2_CONTEXT } from '../../../../lib/graphql/helpers';
-import type { Account, DashboardOrdersQueryVariables } from '../../../../lib/graphql/types/v2/graphql';
-import useQueryFilter from '../../../../lib/hooks/useQueryFilter';
+import IncomingContributionsForHosted from './IncomingContributionsForHosted';
+import IncomingContributionsForOrganizations from './IncomingContributionsForOrganizations';
 
-import { DashboardContext } from '../../DashboardContext';
-import DashboardHeader from '../../DashboardHeader';
-import { childAccountFilter } from '../../filters/ChildAccountFilter';
-import { tierFilter } from '../../filters/TierFilter';
-import type { DashboardSectionProps } from '../../types';
-
-import ContributionsTable from './ContributionsTable';
-import type { FilterMeta as BaseFilterMeta } from './filters';
-import {
-  ContributionAccountingCategoryKinds,
-  filters as baseFilters,
-  schema as baseSchema,
-  toVariables as baseToVariables,
-} from './filters';
-import { PausedIncomingContributionsMessage } from './PausedIncomingContributionsMessage';
-import { dashboardOrdersQuery } from './queries';
-import { useIncomingOutgoingContributionViews } from './views';
-
-const schema = baseSchema.extend({ tier: tierFilter.schema, account: childAccountFilter.schema });
-
-type FilterValues = z.infer<typeof schema>;
-type FilterMeta = BaseFilterMeta & {
-  childrenAccounts?: Account[];
-  selectedAccountSlug?: string;
-};
-
-const toVariables: FiltersToVariables<FilterValues, DashboardOrdersQueryVariables, FilterMeta> = {
-  ...baseToVariables,
-  tier: tierFilter.toVariables,
-  account: (value, key, meta) => {
-    if (meta?.childrenAccounts && !meta.childrenAccounts.length) {
-      return { includeChildrenAccounts: false };
-    } else if (!value) {
-      return { includeChildrenAccounts: true };
-    } else {
-      return { slug: value, includeChildrenAccounts: false };
-    }
-  },
-};
-
-const filters: FilterComponentConfigs<FilterValues, FilterMeta> = {
-  ...baseFilters,
-  tier: tierFilter.filter,
-  account: childAccountFilter.filter,
-};
-
-const IncomingContributions = ({ accountSlug }: DashboardSectionProps) => {
-  const { account } = useContext(DashboardContext);
-
-  const { views, refetch: refetchViews } = useIncomingOutgoingContributionViews(accountSlug, 'INCOMING');
-
-  const filterMeta: FilterMeta = {
-    currency: account.currency,
-    accountSlug: account.slug,
-    childrenAccounts: account.childrenAccounts?.nodes ?? [],
-    hostSlug: account.isHost ? account.slug : undefined,
-    includeUncategorized: true,
-    accountingCategoryKinds: ContributionAccountingCategoryKinds,
-  };
-
-  const queryFilter = useQueryFilter({
-    schema,
-    toVariables,
-    meta: filterMeta,
-    views,
-    filters,
-  });
-
-  const { data, loading, error, refetch } = useQuery(dashboardOrdersQuery, {
-    variables: {
-      slug: accountSlug,
-      filter: 'INCOMING',
-      includeIncognito: true,
-      includeChildrenAccounts: true,
-      ...queryFilter.variables,
-    },
-    context: API_V2_CONTEXT,
-    fetchPolicy: typeof window !== 'undefined' ? 'cache-and-network' : 'cache-first',
-  });
-
-  // Add dynamic values to meta
-  queryFilter.meta = {
-    ...filterMeta,
-    selectedAccountSlug: queryFilter.values.account || undefined,
-  };
-
-  const handleRefetch = React.useCallback(() => {
-    refetch();
-    refetchViews();
-  }, [refetch, refetchViews]);
-
-  const currentViewCount = views.find(v => v.id === queryFilter.activeViewId)?.count;
-  const nbPlaceholders = currentViewCount < queryFilter.values.limit ? currentViewCount : queryFilter.values.limit;
-
-  const orders = data?.account?.orders ?? { nodes: [], totalCount: 0 };
-
-  return (
-    <div className="flex flex-col gap-4">
-      <DashboardHeader
-        title={<FormattedMessage id="IncomingContributions" defaultMessage="Incoming Contributions" />}
-        description={
-          <FormattedMessage id="IncomingContributions.description" defaultMessage="Contributions to your account." />
-        }
-      />
-
-      <PausedIncomingContributionsMessage accountSlug={accountSlug} />
-
-      <ContributionsTable
-        accountSlug={accountSlug}
-        queryFilter={queryFilter}
-        views={views}
-        orders={orders}
-        loading={loading}
-        nbPlaceholders={nbPlaceholders}
-        error={error}
-        refetch={handleRefetch}
-      />
-    </div>
-  );
+const IncomingContributions = ({ account }) => {
+  if (account.type === 'ORGANIZATION') {
+    return <IncomingContributionsForOrganizations accountSlug={account.slug} />;
+  } else {
+    return <IncomingContributionsForHosted accountSlug={account.slug} />;
+  }
 };
 
 export default IncomingContributions;
