+++ lang/ca.json
@@ -127,6 +127,7 @@
   "2f9kCJ": "Expected date for funds received",
   "2fKAKF": "I don't want to contribute to Open Collective",
   "2FxRMv": "<Individual></Individual> added <FromAccount></FromAccount> as an <MemberRole></MemberRole> for <Account></Account>",
+  "2IsUyp": "Contributions made to your Organization.",
   "2JPEip": "Transfer funds between your accounts.",
   "2KTLpo": "Unhost",
   "2kY5p5": "Total amount received this period",
@@ -2465,7 +2466,6 @@
   "IYoPSx": "Fund's overview",
   "IzFWHI": "Chart of Accounts",
   "izhuHE": "Bill To",
-  "IZotDb": "Contributions made to your organization and Collectives you host.",
   "j/a0gE": "The \"{planTitle}\" plan gives you access with limited usage mentioned below. Any usage exceeding the plan limits will incur additional charges.",
   "J/Pgyh": "A 3 letter identifier of the host currency.",
   "j0+O6x": "You'll be redirected shortly.",
@@ -4108,6 +4108,7 @@
   "TnG9DJ": "Create grant request",
   "tnRDuU": "Revoke",
   "To": "To",
+  "To33FZ": "Contributions made to your Organization and Collectives you host.",
   "ToCollectives": "To Collectives",
   "TodaysCharge": "Today's charge",
   "ToDo.EmptyStateMessage": "You’re all caught up.",
