+++ lang/de.json
@@ -127,6 +127,7 @@
   "2f9kCJ": "Voraussichtliches Datum für den Geldeingang",
   "2fKAKF": "Ich möchte nicht zu Open Collective beitragen",
   "2FxRMv": "<Individual></Individual> added <FromAccount></FromAccount> as an <MemberRole></MemberRole> for <Account></Account>",
+  "2IsUyp": "Contributions made to your Organization.",
   "2JPEip": "Übertagen Sie Geld zwischen Ihren Konten.",
   "2KTLpo": "Moderation entfernen",
   "2kY5p5": "Erhaltener Gesamtbetrag in diesem Zeitraum",
@@ -2465,7 +2466,6 @@
   "IYoPSx": "Fund's overview",
   "IzFWHI": "Kontenplan",
   "izhuHE": "Rechnung an",
-  "IZotDb": "Contributions made to your organization and Collectives you host.",
   "j/a0gE": "The \"{planTitle}\" plan gives you access with limited usage mentioned below. Any usage exceeding the plan limits will incur additional charges.",
   "J/Pgyh": "A 3 letter identifier of the host currency.",
   "j0+O6x": "You'll be redirected shortly.",
@@ -4108,6 +4108,7 @@
   "TnG9DJ": "Create grant request",
   "tnRDuU": "Revoke",
   "To": "An",
+  "To33FZ": "Contributions made to your Organization and Collectives you host.",
   "ToCollectives": "To Collectives",
   "TodaysCharge": "Today's charge",
   "ToDo.EmptyStateMessage": "You’re all caught up.",
