+++ lang/es.json
@@ -127,6 +127,7 @@
   "2f9kCJ": "Fecha prevista para recibir los fondos",
   "2fKAKF": "No quiero hacer una contribución a Open Collective",
   "2FxRMv": "<Individual></Individual> added <FromAccount></FromAccount> as an <MemberRole></MemberRole> for <Account></Account>",
+  "2IsUyp": "Contributions made to your Organization.",
   "2JPEip": "Transfiere fondos entre tus cuentas.",
   "2KTLpo": "Eliminar Anfitrión",
   "2kY5p5": "Importe total recibido este periodo",
@@ -2465,7 +2466,6 @@
   "IYoPSx": "Resumen del fondo",
   "IzFWHI": "Plan de Cuentas",
   "izhuHE": "Facturar a",
-  "IZotDb": "Contributions made to your organization and Collectives you host.",
   "j/a0gE": "The \"{planTitle}\" plan gives you access with limited usage mentioned below. Any usage exceeding the plan limits will incur additional charges.",
   "J/Pgyh": "Identificador de 3 letras de la moneda del Anfitrión.",
   "j0+O6x": "Serás redirigido en breve.",
@@ -4108,6 +4108,7 @@
   "TnG9DJ": "Crear solicitud de subvención",
   "tnRDuU": "Revoke",
   "To": "A",
+  "To33FZ": "Contributions made to your Organization and Collectives you host.",
   "ToCollectives": "Para Colectivos",
   "TodaysCharge": "El cargo de hoy",
   "ToDo.EmptyStateMessage": "You’re all caught up.",
