+++ lang/fr.json
@@ -127,6 +127,7 @@
   "2f9kCJ": "Date prévue de réception des fonds",
   "2fKAKF": "Je ne veux pas contribuer à Open Collective",
   "2FxRMv": "<Individual></Individual> added <FromAccount></FromAccount> as an <MemberRole></MemberRole> for <Account></Account>",
+  "2IsUyp": "Contributions made to your Organization.",
   "2JPEip": "Transférer des fonds entre vos comptes.",
   "2KTLpo": "Non hébergé",
   "2kY5p5": "Montant total reçu sur cette période",
@@ -2465,7 +2466,6 @@
   "IYoPSx": "Aperçu du Fonds",
   "IzFWHI": "Plan Comptable",
   "izhuHE": "Facturer à",
-  "IZotDb": "Contributions made to your organization and Collectives you host.",
   "j/a0gE": "The \"{planTitle}\" plan gives you access with limited usage mentioned below. Any usage exceeding the plan limits will incur additional charges.",
   "J/Pgyh": "Un identifiant de 3 lettres de la devise de l'Hôte.",
   "j0+O6x": "Vous allez être redirigé rapidement.",
@@ -4108,6 +4108,7 @@
   "TnG9DJ": "Créer une demande de subvention",
   "tnRDuU": "Revoke",
   "To": "À",
+  "To33FZ": "Contributions made to your Organization and Collectives you host.",
   "ToCollectives": "Aux Collectifs",
   "TodaysCharge": "Sous-total",
   "ToDo.EmptyStateMessage": "Vous êtes à jour.",
