+++ lang/pt-BR.json
@@ -127,6 +127,7 @@
   "2f9kCJ": "Data prevista para fundos recebidos",
   "2fKAKF": "Não quero contribuir com a Open Collective",
   "2FxRMv": "<Individual></Individual> added <FromAccount></FromAccount> as an <MemberRole></MemberRole> for <Account></Account>",
+  "2IsUyp": "Contributions made to your Organization.",
   "2JPEip": "Transferir valores entre suas contas.",
   "2KTLpo": "Remover anfitrião",
   "2kY5p5": "Valor total recebido neste período",
@@ -2465,7 +2466,6 @@
   "IYoPSx": "Visão geral do fundo",
   "IzFWHI": "Plano de Contas",
   "izhuHE": "Cobrar para",
-  "IZotDb": "Contributions made to your organization and Collectives you host.",
   "j/a0gE": "The \"{planTitle}\" plan gives you access with limited usage mentioned below. Any usage exceeding the plan limits will incur additional charges.",
   "J/Pgyh": "Um identificador de 3 letras da moeda do host.",
   "j0+O6x": "Você será redirecionado em breve.",
@@ -4108,6 +4108,7 @@
   "TnG9DJ": "Criar solicitação de subsídio",
   "tnRDuU": "Revoke",
   "To": "Para",
+  "To33FZ": "Contributions made to your Organization and Collectives you host.",
   "ToCollectives": "Para Coletivos",
   "TodaysCharge": "Taxa de hoje",
   "ToDo.EmptyStateMessage": "Vocês estão todos atualizados.",
