+++ lang/zh.json
@@ -127,6 +127,7 @@
   "2f9kCJ": "预计收到资金的日期",
   "2fKAKF": "我不想为 Open Collective 做出贡献",
   "2FxRMv": "<Individual></Individual> 为 <Account></Account> 添加了<MemberRole></MemberRole> <FromAccount></FromAccount>",
+  "2IsUyp": "Contributions made to your Organization.",
   "2JPEip": "账户间资金调动。",
   "2KTLpo": "取消托管",
   "2kY5p5": "本期收到的总金额",
@@ -2465,7 +2466,6 @@
   "IYoPSx": "Fund's overview",
   "IzFWHI": "Chart of Accounts",
   "izhuHE": "付给",
-  "IZotDb": "Contributions made to your organization and Collectives you host.",
   "j/a0gE": "The \"{planTitle}\" plan gives you access with limited usage mentioned below. Any usage exceeding the plan limits will incur additional charges.",
   "J/Pgyh": "A 3 letter identifier of the host currency.",
   "j0+O6x": "You'll be redirected shortly.",
@@ -4108,6 +4108,7 @@
   "TnG9DJ": "Create grant request",
   "tnRDuU": "Revoke",
   "To": "发送至",
+  "To33FZ": "Contributions made to your Organization and Collectives you host.",
   "ToCollectives": "To Collectives",
   "TodaysCharge": "今日收入",
   "ToDo.EmptyStateMessage": "You’re all caught up.",
